requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.5.1
_libavif_api=1.3.0=h57928b3_2
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.13.3=pypi_0
aiosignal=1.4.0=pypi_0
akshare=1.18.9=pypi_0
aom=3.9.1=he0c23c2_0
async-timeout=5.0.1=pypi_0
attrs=25.4.0=pypi_0
beautifulsoup4=4.14.3=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.12.2=haa95532_0
cairo=1.18.4=he9e932c_0
certifi=2026.1.4=pypi_0
cffi=2.0.0=pypi_0
charset-normalizer=3.4.4=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.0.5=py39h59b6b97_0
curl-cffi=0.13.0=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
dav1d=1.2.1=hcfcfb64_0
decorator=5.2.1=pypi_0
eigen=3.4.0=h477610d_1
et-xmlfile=2.0.0=pypi_0
expat=2.7.3=h885b0b7_4
ffmpeg=6.1.1=haa678a1_5
filelock=3.19.1=pypi_0
fontconfig=2.15.0=hd211d86_0
fonttools=4.60.1=py39h02ab6af_0
freeglut=3.8.0=hfcef157_0
freetype=2.13.3=h0620614_0
fribidi=1.0.16=hfd05255_0
frozenlist=1.8.0=pypi_0
fsspec=2025.10.0=pypi_0
future=1.0.0=pypi_0
gflags=2.2.2=he0c23c2_1005
glog=0.5.0=h4797de2_0
graphite2=1.3.14=hd77b12b_1
gst-plugins-base=1.24.12=h91a6125_1
gstreamer=1.24.12=hfb93a4f_1
gstreamer-orc=0.4.41=h1f81b68_0
harfbuzz=10.2.0=he2f9f60_1
hdf5=1.14.5=ha36df97_2
html5lib=1.1=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=73.1=h6c2663c_0
idna=3.11=pypi_0
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.6=pypi_0
joblib=1.5.3=pypi_0
jpeg=9f=ha349fce_0
jsonpath=0.82.2=pypi_0
keyboard=0.13.5=pypi_0
kiwisolver=1.4.4=py39hd77b12b_0
lcms2=2.17=h3732fa5_0
lerc=4.0.0=h6470a55_1
libabseil=20250127.0=cxx17_h4eb7d71_0
libavif=1.3.0=he916da2_2
libavif16=1.3.0=he916da2_2
libdeflate=1.22=h2466b09_0
libexpat=2.7.3=h885b0b7_4
libffi=3.4.4=hd77b12b_1
libglib=2.84.4=hfaec014_0
libiconv=1.16=h2bbff1b_3
libkrb5=1.21.3=h885b0b7_4
libogg=1.3.5=h2466b09_1
libopenjpeg=2.5.4=h02ab6af_1
libpng=1.6.50=h46444df_0
libpq=17.6=h652a1e2_0
libprotobuf=5.29.3=he9d8c4a_0
libtheora=1.2.0=habf309d_1
libtiff=4.7.1=h3a18249_0
libvorbis=1.3.7=h5112557_2
libwebp-base=1.6.0=hbf3958f_0
libxml2=2.13.9=h6201b9f_0
libzlib=1.3.1=h02ab6af_0
lxml=6.0.2=pypi_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.5=pypi_0
matplotlib=3.6.2=py39haa95532_0
matplotlib-base=3.6.2=py39h1094b8e_0
mini-racer=0.12.4=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h827c3e9_2
mkl_fft=1.3.11=py39h827c3e9_0
mkl_random=1.2.8=py39hc64d2fc_0
mpmath=1.3.0=pypi_0
multidict=6.7.0=pypi_0
mysql-common=9.3.0=hc394544_4
mysql-libs=9.3.0=h58c8544_4
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=pypi_0
numpy=1.26.4=pypi_0
opencv=4.12.0=py39h6fcf9f4_0
openjpeg=2.5.4=h24db6dd_0
openpyxl=3.1.5=pypi_0
openssl=3.6.0=h725018a_0
packaging=25.0=py39haa95532_1
pandas=2.3.3=pypi_0
pcre2=10.46=h5740b90_0
pillow=11.3.0=py39hb328d1f_0
pip=25.3=pyhc872135_0
pixman=0.46.4=h4043f72_0
propcache=0.4.1=pypi_0
pycparser=2.23=pypi_0
pygame=2.6.1=pypi_0
pynput=1.8.1=pypi_0
pyparsing=3.2.0=py39haa95532_0
pypiwin32=223=pypi_0
pyqt=6.9.1=py39h12ec796_0
pyqt5=5.15.11=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.17.1=pypi_0
pyqt6-sip=13.10.2=py39h630b2a1_0
pyswarm=0.6=pypi_0
pyswarms=1.3.0=pypi_0
python=3.9.25=h716150d_1
python-dateutil=2.9.0post0=py39haa95532_2
pytz=2025.2=pypi_0
pywin32=311=pypi_0
pyyaml=6.0.3=pypi_0
qtbase=6.9.2=h06bae2a_5
qtdeclarative=6.9.2=h88b4c33_1
qtsvg=6.9.2=h30ace32_1
qttools=6.9.2=h7e7b719_1
qtwebchannel=6.9.2=heb02b0b_1
qtwebsockets=6.9.2=heb02b0b_1
rav1e=0.7.1=ha073cba_3
requests=2.32.5=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.13.1=pypi_0
setuptools=80.9.0=py39haa95532_0
sip=6.12.0=py39h706e071_0
six=1.17.0=py39haa95532_0
soupsieve=2.8.1=pypi_0
sqlite=3.51.0=hda9a48d_0
svt-av1=3.1.2=hac47afa_0
sympy=1.14.0=pypi_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.15=hf199647_0
tomli=2.2.1=py39haa95532_0
torch=2.7.0+cu128=pypi_0
torchaudio=2.7.0+cu128=pypi_0
torchvision=0.22.0+cu128=pypi_0
tornado=6.5.1=py39h827c3e9_0
tqdm=4.67.1=pypi_0
typing-extensions=4.15.0=pypi_0
tzdata=2025.3=pypi_0
ucrt=10.0.22621.0=haa95532_0
urllib3=2.6.3=pypi_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
webencodings=0.5.1=pypi_0
wheel=0.45.1=py39haa95532_0
win10toast-click=0.1.2=pypi_0
winotify=1.1.0=pypi_0
xlrd=2.0.2=pypi_0
xz=5.6.4=h4754444_1
yarl=1.22.0=pypi_0
zlib=1.3.1=h02ab6af_0
zstd=1.5.7=h56299aa_0
